Revert "Update: add option to translate story only (#310)"
This reverts commit c00b4054ea0aeceea1b5d2ac99ef213c62715032.
