+++ data/static/style/BLHXFY.css
@@ -730,9 +730,3 @@ body .cnt-raid .prt-raid-log {
 .cnt-detail #box-sub-blhxfy {
   top: 248px;
 }
-
-.block-story-only {
-  position: absolute;
-  top: 100px;
-  left: 5px;
-}
