+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "blhxfy",
-  "version": "1.11.0",
+  "version": "1.10.14",
   "description": "碧蓝幻想汉化脚本",
   "main": "./src/main.js",
   "scripts": {
