+++ src/setting/hideSidebar.js
@@ -1,0 +1,10 @@
+import config from '../config'
+import { insertCSS } from '../fetch'
+
+const hideSidebar = () => {
+  if (config.hideSidebar) {
+    insertCSS('hide-sidebar')
+  }
+}
+
+hideSidebar()
