+++ src/setting/keepBgm.js
@@ -1,0 +1,11 @@
+import config from '../config'
+
+const keepBgm = () => {
+  if (config.keepBgm) {
+    window.addEventListener('blur', function (e) {
+      e.stopImmediatePropagation()
+    }, false)
+  }
+}
+
+keepBgm()
